refactor: promotion 관련 유틸 함수 도메인 폴더로 분리
