+++ frontend/src/pages/PromotionPage/components/list/PromotionCard/PromotionCard.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import { USER_EVENT } from '@/constants/eventName';
 import useMixpanelTrack from '@/hooks/Mixpanel/useMixpanelTrack';
 import { PromotionArticle } from '@/types/promotion';
-import { getDDay } from '@/utils/getDday';
+import { getDDay } from '@/pages/PromotionPage/utils/getDday';
 import CardMeta from './CardMeta/CardMeta';
 import ClubTag from './ClubTag/ClubTag';
 import DdayBadge from './DdayBadge/DdayBadge';
