+++ frontend/src/pages/PromotionPage/utils/getDday.test.ts
