+++ frontend/src/pages/PromotionPage/utils/getDday.ts
