+++ frontend/src/pages/PromotionPage/utils/promotionNotification.ts
